Fixed implicit cast from int to float.

+++ src/Kernels.cpp
@@ -60,7 +60,7 @@ void InitKernel(double norm)
 		for (int j = 0; j < P.NCP; j++)
 		{
 			cell *m = CellLookup[j];
-			l->tot_prob += (l->max_trans[j] = (float)numKernel(dist2_cc_min(l, m))) * m->n;
+			l->tot_prob += (l->max_trans[j] = (float)numKernel(dist2_cc_min(l, m))) * (float)m->n;
 		}
 	}
 }
